backend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
sqlalchemy==1.4.23
flask-jwt-extended==4.3.1
python-dotenv==0.19.0
numpy>=1.22.0,<1.24.0
Pillow==8.3.2
opencv-python-headless==4.8.0.76
tensorflow==2.12.0
# Using deepface as an alternative to face_recognition
deepface==0.0.75
bcrypt==4.0.1
psycopg2-binary==2.9.7
requests==2.31.0